+++ Controller/Adminhtml/Sales/Order/CancelSubscription.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Nuvei\Checkout\Controller\Adminhtml\Sales\Order;
+
+/**
+ * @author Nuvei
+ */
+class CancelSubscription extends \Magento\Backend\App\Action
+{
+    public function execute()
+    {
+        die('test');
+    }
+    
+}